TTLHashSet.Add() sets an expiry time on the key when adding it
This is so that we don't indefinitely view URLs we've crawled as being
seen. A 5 hour expiry time is liberal enough given our current page
turn around time for various pages.

This might need some extra tweaking for things like FCO Travel pages.

Alternatively, we could in the future follow Cache-Control headers for
pages we crawl instead of picking an arbitrary (fixed) time.