When working with queues, it can successfully consume/publish
This is the meat of our worker. Once the relevant exchanges and queues
have been created and binded we should be able to consume them to
complete whatever work we need to do.

Additionally, when we see further work to be completed we should be
able to add it to the back of the queue.