ExtractURLs is a helper function to extract URLs from (DOM) responses
This helper method will provide us with a set of URLs that have been
found so that we can add them to the back of the queue for crawling as
we pass over each of the URLs we've found.

In addition to extract all `a[@href]` URLs it will extract:

 - `img[@src]`
 - `link[@href]`
 - `script[@src]`

It doesn't provide any filtering of these URLs. This logic should live
elsewhere.