Reduce the publish channel buffer and acknowledge last
On average we pick up between 50 and 100 URLs per page. If the process
were to die we'd end up losing 1000 URLs and wouldn't crawl them
because we're acknowledging the message up front.

Acknowledge the message last and only keep a buffer of up to 100 URLs
that we want to publish to the exchange.