Delete mirror root after tests run
Delete the mirror root directory, as defined by the `MIRROR_ROOR`
environment variable, after each of the tests has run.

Do this after _each_ test in case so as not to cross-contaminate
fixtures between tests.

Also add a note to the README to warn that this will happen.

Add a new file, `helpers_test.go` to contain the shared function that
deletes the mirror root. This would have been named `test_helpers.go`,
but it seems Ginkgo requires it to be named with the `_test.go` suffix.

I also wanted to raise an error and call `panic()` or `os.Exit()` if
`MIRROR_ROOT` already exists on disk to prevent someone from deleting
files they want to keep but I couldn't get Ginkgo to exit before it runs
AfterEach(), which deletes those files.