Write content to disk in separate goroutines
Move the `WriteToDisk()` method from the `CrawlerMessageItem` interface
to run as part of the main workflow in `workflow.go` because writing is
a separate concern to crawling.

Update the tests accordingly and improve them so that we check for
existence of the file created on disk.

Use two goroutines for symmetry with `CrawlURL()`. Note that we reject
the message without requeueing it as a failure to write to disk is
likely to require some manual intervention and I don't think it's wise
for us to retry.