Explicitly declare channels for readability
Declare the variables we use as
channels before they are used to improve readability.

When I first looked at this code, being new to Go, it took me a moment
to realise that `acknowledge`, `crawlItems` and `extract` are in fact
channels. By using `var` instead of the `:=` shortcut, it should be
clearer what these are.

I also considered (credit to @dancarley) creating a custom type such as:
```go
type itemChan <-chan *CrawlerMessageItem
```

However, I think using this and replacing all instances of `<-chan
*CrawlerMessageItem` in the codebase, while reducing duplication, would
obscure its meaning.