Add hostname to user agent
So that we know which server the crawler is running on.

Only check for the string "GOV.UK Crawler Worker" in our tests as it's
not easy for us to mock the system hostname and testing against
`os.Hostname()` would be of minimal benefit.

Also removes the function that trims the body of the HTML response, as
there's no reason for there to be spaces around the text in the response
body.