Symlink self into third_party/src
So that we don't clone a new copy of our own code and ignore local changes
when building/testing. This is what most of the CoreOS projects (where
third_party.go originated from) do:

https://github.com/coreos/etcd/blob/v0.4.4/build#L3-6

Should fix a problem that @rjw1 and @mattbostock experienced.

make(1) won't re-run the commands if the symlink already exists because the
target name matches the "file". However you will need to delete your
existing directory if you've already used make in the past.