Test the content that is written to disk
In the test for WriteItemToDisk(), we were previously only checking that
the file existed on disk.

Expand this test to check the contents of the file written to disk. The
file's existence is confirmed by the fact of it having content and we
also check for errors when reading the file.