Disable noWait when deleting queues/exchanges
The use of noWait when deleting queues and exchanges in tests means that we
don't catch any errors (they go to the notifyClose channel which is handled
within QueueConnection) and that their presence can affect the behaviours of
tests in other describe blocks.

This now causes the tests to fail when deleting the exchanges. Which we'll
need to fix before merging this commit:

    Expected
        <*amqp.Error | 0xc2101734e0>: {
            Code: 503,
            Reason: "COMMAND_INVALID - unimplemented method",
            Server: true,
            Recover: false,
        }
    to be nil