[#73135254] HealthCheck is an object that can return an HTTP Handler
To build out our HTTP health check endpoint we're going to convert our
simple function into an object with multiple methods attached. These
methods return the Status() and an HTTPHandler(). The latter can be
used as part of Go's HTTP server and attached using
http.HandlerFunc().

This is wired into the main function and the port can be configured
using environment variables.