Extract log.Fatal from queue notifyClose
Extract the log.Fatal() call, which happens when QueueConnection gets an
amqp.ErrClosed error from the NotifyClose channel, into a struct field.

This will allow us to inject our own non-fatal function during tests to
assert that it would exit under the right conditions.

I've made it a struct field rather than a package variable so that we don't
have to worry about resetting it back again after the test.