Failing test for exit on amqp.ErrClosed
QueueConnection looks like it should cause the crawler worker to exit if it
encounters a closed AMQP channel/connection. However it doesn't look like
this is working, as shown by the timeout when waiting for the error on the
fatalErrs channel.

My understanding of the docs is that the NotifyClose channel will be closed
instead of actually receiving amqp.ErrClosed:

http://godoc.org/github.com/streadway/amqp#Channel.NotifyClose

> The chan provided will be closed when the Channel is closed and on a
> graceful close, no error will be sent.