Use publisher/producer to delete queues
To avoid us getting spurious NotifyClose errors when we delete the queue
without calling Cancel() or Close().

This highlighted the need to Ack() any messages that we receive on the
consumer channel. If we don't do this, then when we call consumer.Close()
the un-acked messages are returned to the queue and QueueDelete() returns a
non-zero value.