Exit on non-recoverable AMQP errors
Abort the crawler if we encounter any AMQP errors that are non-recoverable
and would cause all subsequent operations to receive a "channel/connection
is not open" error.

The story that I was originally working on was to reconnect if we lost
connection to the AMQP server. But this turned out to be more complex and
not as useful as we'd originally hoped, given that we can't meaningfully
continue without the queue (unlike Redis). So now it does the thing we
originally intended and we'll rely on Upstart or Puppet restart the process
for us - soon to be documented by Kush.