Return an error if redirect status encountered
Cause `http_crawler.Crawl()` to return an error if a HTTP redirect
status code is encountered.

This will be caught by `CrawlURL()` in `workflow.go`[1] such that the
contents of this URL will not be written to disk.

To prevent the redirect, we specify a custom HTTP client that defines a
`CheckRedirect` function. Go's default `CheckRedirect` function follows
up to 10 redirects[2].

I looked at using `transport.RoundTrip()`[3], which results in a single
transaction (thus redirects aren't followed automatically), but
preferred to keep `http.Client` so that we can use its high-level
features.

[1]:
https://github.com/alphagov/govuk_crawler_worker/blob/2e4e1cc52796d3507e4abd635771346a9acdc519/workflow.go#L62-L65
[2]: http://golang.org/src/pkg/net/http/client.go?s=882:2405#L389
[3]: https://groups.google.com/forum/#!topic/golang-nuts/PmYm3h2J3FE