Move RetryStatusCodes() further down
It's less important than the functions that followed it; move it further
down the file so that the code is easier to read (most important
functions first).

I had wanted to make this function private as it's not used outside of
`http_crawler` in the main code, but it turns out that the Ginkgo tests
fail if it's defined as a private method. Usually we wouldn't want to
test private methods, but in this case I think this test adds value.