Track the number of URLs we still have to publish
So that we can better understand one of the failure cases (early exit)
we should track the number of URLs that we haven't published back to
the exchange and that we have in progress.

Switched from the Etsy example Go code to another StatsD library
because the original Etsy code didn't support sending gauges.