[#73135884] Increase back off time for 429 response codes
Increase the amount of time we back off from crawling when we receive
a 429 HTTP status code. Otherwise we won't have waited long enough for
the server to let us continue.

Also remove back off times for 5XX responses as we'll want to count
the number of times we receive a 5XX response and then manage how to
reject the message in the future.