Test that we won't keep retrying URL indefinitely
Add a test to verify that we don't keep trying to crawl an item that
repeatedly returns HTTP 500 errors; in this case we allow 4 retries.

Because nothing should be sent over the `crawled` channel  (since the
`*CrawlMessageItem` being processed will be rejected by `CrawlURL()`),
we use the value stored in the Redis-backed `*TTLHashSet` to determine
if the maximum number of retires has been reached.