Check URL is blacklisted when reading from queue
Check if a URL is blacklisted as we read it from the RabbitMQ queue to
prevent us from crawling blacklisted URLs.

Previously we would only check URLs that were extracted from a crawled
HTML document.

Note that in the tests, we inspect the RabbitMQ queue to verify that the
message has been rejected. This is necessary because the channel
returned by `ReadFromQueue()` will always have a zero length (since the
message is rejected), so we can't perform any assertions on that
channel.