Use github.com/golang/glog for logging
Use Google's glog package for logging, which enables us to log at
varying levels of severity:

`INFO`, `WARNING`, `ERROR` and `FATAL`

glog is a pure implementation of leveled logs based on Google's C++
package:
http://code.google.com/p/google-glog

The defaults output to temporary files; this can be changed by setting
the `-logtostderr` flag to true.  The default output shows ERROR and
FATAL messages only; this can be changed by setting the
`-stderrthreshold` flag[1].

This package also provides the `-log_backtrace_at` flag[2], which will
cause a backtrace to be output when execution reaches a logging
statement at the location specified in that flag.

[1]:
https://github.com/golang/glog/blob/d1c4472bf2efd3826f2b5bdcc02d8416798d678c/glog.go#L46
[2]:
https://github.com/golang/glog/blob/d1c4472bf2efd3826f2b5bdcc02d8416798d678c/glog.go#L55-L61