Set Content-Type HTTP header and remove `expectedFilePath` variable
Update these tests for the `item.RelativeFilePath()` so that they
still behave as we expect before we start to manipulate the underlying
code to support multiple file formats.

Also removed the `expectedFilePath` variable as it was confusing to
follow. It didn't make sense to keep on changing this single variable
when the reference could just as easily be kept inlined. This is now
the case so the tests are more self contained.