Use an alphagov fork of the redis library
In the last 24 hours a bug was added to the Redis library that we're
using. The bug has been tracked[1] and a small test case[2] has been
written to reproduce it.

In the mean time, we're going to use a fork of the radix Redis library
that we pin the version for instead of directly checking the library
into a `third_party` directory using `third_party.go`.

This was discussed with @bradleywright as the right thing to do short
term until we can get around to fixing the library in question.

[1] https://github.com/fzzy/radix/issues/35
[2]
https://github.com/KushalP/radix/compare/try-to-reproduce-bug-as-tests-added-in-35

P.S. We still haven't defined a consistent way of managing
dependencies across alphagov projects that have been written in Go.
This was deemed a good enough solution to unblock us for now.