Retrieve deps for subdirectories and only download
Previously, we were retrieving dependencies for subdirectories but for
an unknown reason `go get` would not retrieve dependencies required by
`_test.go` files in subdirectories. We're not actually using any such
dependencies currently that aren't already required by the files we are
testing, but I came across the issue when writing an provisional test.

Also, only download the dependencies but don't install them. This is
necessary to prevent the following errors when using the `./...`
wildcard in conjunction with `third_party.go`:

    » make
    [11:44:53]
    rm -rf third_party/pkg/*/github.com/alphagov/govuk_crawler_worker
    go run third_party.go get -t -v ./...
    go install: no install location for directory
    /Users/matt/govuk/govuk_crawler_worker/http_crawler outside GOPATH
    go install: no install location for directory
    /Users/matt/govuk/govuk_crawler_worker/queue outside GOPATH
    go install: no install location for directory
    /Users/matt/govuk/govuk_crawler_worker/ttl_hash_set outside GOPATH
    go install: no install location for directory
    /Users/matt/govuk/govuk_crawler_worker/util outside GOPATH
    exit status 1
    make: *** [deps] Error 1