Exit if the RabbitMQ channel closes
We currently exit if the RabbitMQ connection drops, the Crawler simply
logs a fatal error and exits.

This commit does the same for when the RabbitMQ channel (NB channel, not
connection) drops.

Given the nature of the work the Crawler does, this does not present an
issue and it will continue where it left off when restarted. With our
current configuration, the service will be automatically restarted
within 30 minutes if it fails.

If this happens too frequently, we may consider improving this. But
currently, I have not yet seen an example of the channel closing without
the connection dropping at the same time.