Make INFO the default log level
Make `INFO` the default log level and remove the `-debug` flag.

Previously, the `INFO` level was very noisy and so it was undesirable to
use this as the default. In the previous commit, many noisy log messages
were demoted to `DEBUG` level, which makes `INFO` much more usable.

As a consequence, the `-debug` flag is no longer necessary. `-verbose`
will now show `DEBUG` messages and `INFO` log messages will be shown by
default.

Notable examples of `INFO` messages include the AMQP and Redis
connection messages shown when the application initialises.