Rename queue.QueueConnection to queue.Connection
Similar to 6d64219c, rename the struct to make it less awkward when
qualified with its package name, `queue`.

Renaming this causes some ambiguity in `workflow_test.go` because the
`queue` package's exported identifiers have been imported into the test
source file and, as such, are accessed without the qualifying `queue`
prefix. But I think that's a smell that we should fix at some point.