Sleep for 5 seconds on HTTP 5XX error
Previously, we only backed off when a HTTP 429 (Too Many Requests) error
was encountered.

Extend this so that we also back off when we encounter a HTTP 5XX error,
which will be returned when an application is unable to serve a
response.

Note that we only increment the retry count on `ttlHashSet` for 5XX
errors, since a HTTP 429 is just an indication that we are crawling too
much too quickly, rather than an actual error in serving the content. As
such, it should be safe to retry on HTTP 429 errors until we have a
successful response.