Use new Logrus hook to send exceptions to Errbit
The previous hook used to send exceptions to Errbit's
Airbrake-compatible API was raising errors because we weren't adding an
'error' field to each log entry:

    {"endpoint":"https://errbit.preview.alphagov.co.uk","level":"warning","msg":"Exceptions sent to Airbrake must have an 'error' key with the error","source":"airbrake","time":"2015-03-09T11:20:55Z"}

I wrote an integration test for the previous hook which documents this
behaviour:
https://github.com/Sirupsen/logrus/pull/145/files#diff-b9bcfed63bb364328842ba8e05acaa26R46

The previous hook is useful if you need to pass an error type to Errbit,
but in our case we only need to send the error message. So I raised a PR
upstream to implement a new hook:
https://github.com/Sirupsen/logrus/pull/146

That PR has not yet been merged upstream, so this commit swaps out the
old hook for the new one in our `alphagov/logrus` fork.