Fix dependencies
These dependencies were no longer available or functional:

- code.google.com has shutdown and the commitish referenced for the
  `cascadia` package is not in the tree of the GitHub repo it has moved
  to

- the `atom` package has moved to golang.org and the commitsh referenced
  does exist in the tree on golang.org

- `goquery` was referencing the old location for the `cascadia` package

Where the old commitish still existed, I've updated these packages only
as far as is necessary for them to work (as it the case for `goquery`):

https://github.com/PuerkitoBio/goquery/compare/ffa8a68c9f2882d21c77b65737c84e2f8690fa1f...d9b2b5ed51f9567423662ac2c64a591a88dd0fce

For the other packages where the commitish we were using no longer
exists, I've used the latest commit.

As it's difficult to verify these changes easily (since we're not
vendoring our code), I'm relying on our tests to verify that nothing has
broken.