Reject ATOM feeds
We don't want the crawler to crawl RSS or ATOM feeds because we don't
want to serve old or outdated copies of feeds to feed readers when
the site is offline.

It's better for feed readers to receive 5xx responses to indicate
things are broken. They'll retry updating feeds later.