Make reconnection tests pass under Go 1.7
These tests were failing due to a change in behaviour on closed
connection handling between 1.5 and 1.6+ (see PoC at
https://gist.github.com/timblair/2578882d17b294b534a24bbd7c4e123e):
we were relying on unspecified behaviour in the runtime which has
since changed.

The test now checks the memory address of the most recent connection in
the underlying ProxyTCP value, which is now exposed via
ProxyTCP.Connections, and should change when the connections are killed.

This is definitely not the most pleasant way of validating the behaviour
(there should be no reason to expose the underlying connection slice
publicly), but given that this type only exists as a test helper, it's
a reasonably pragmatic approach.