Stop decoding URI's before requesting them.
This is converting valid URI's into invalid ones by introducing UTF-8
characters into them.

This changes the behaviour so that we request valid URI's and then we
decode them when the files are saved to disk.

The prompt for this change was many (~4000 a day) errors logged in Whitehall
that originated from UTF-8 requests and Ruby Unicorns inconsistent
handling of them. See https://github.com/alphagov/slimmer/pull/182 for
more info.