Stop sending errors to Airbrake
We are retiring our Errbit instances and moving to Sentry. This commit
stops the crawler from sending errors to Errbit.

The crawler currently sends thousands of errors to Errbit. Because most
are routine exceptions encountered when crawling GOV.UK, these are not
monitored and not actioned. The errors are also logged to Kibana so we
won't lose data.