Reduce redis memory footprint
Now that we are storing the keys in redis we are running into memory issues on
the machine.

We want to stop duplicates from existing in RabbitMQ and we will do this using
redis to store URLs which exist in RabbitMQ, however not all URL are created
equal, so instead of storing all URLs in redis we will try and keep track of
the most common URLs, stopping them being duplicated in rabbit. This means
we will reintroduce some duplication in rabbit, but allow us to manage our redis
memory usage.

We will do this by reducing the redis TTL, if the URL is encountered a again
before the TTL expires, the TTL will be extended, if it only encounters the URL
again after the TTL has expired the url will be inserted into rabbit a 2nd time
and a new TTL set.