Reduce default number of retries to 1
If origin responds with a 500 on the first try and the first retry, it's unlikely that it will start responding with a 200 on any of the 2nd-4th retries.

Retrying 4 times means we send 4x more issues to Sentry than necessary, which creates extra noise and counts against our quota.

Although this is configurable via an environment variable, we don't currently use this functionality (either in Puppet or anywhere else across alphagov).